third_party/blink/web_tests/platform/generic/media/media-fragments/TC0002-expected.txt


Title: TC0002
Fragment: 't=3,3'
Comment: Invalid semantics: start must be smaller than end. The media fragment is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
